+++ src/main/java/org/example/stockitbe/store/order/batch/model/enums/StoreOrderBatchScope.java
@@ -1,0 +1,6 @@
+package org.example.stockitbe.store.order.batch.model.enums;
+
+public enum StoreOrderBatchScope {
+    ALL,
+    STORE
+}
