+++ src/main/java/org/example/stockitbe/store/order/batch/model/enums/StoreOrderBatchTriggerType.java
@@ -1,7 +1,6 @@
-package org.example.stockitbe.store.order.model;
+package org.example.stockitbe.store.order.batch.model.enums;
 
 public enum StoreOrderBatchTriggerType {
     AUTO,
     MANUAL
 }
-
