--- src/main/java/org/example/stockitbe/store/order/model/StoreOrderBatchScope.java
@@ -1,7 +1,0 @@
-package org.example.stockitbe.store.order.model;
-
-public enum StoreOrderBatchScope {
-    ALL,
-    STORE
-}
-
