+++ src/main/java/org/example/stockitbe/store/order/model/dto/StoreOrderDto.java
@@ -8,8 +8,8 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.example.stockitbe.store.order.model.StoreOrderHistoryType;
-import org.example.stockitbe.store.order.model.StoreOrderStatus;
+import org.example.stockitbe.store.order.model.enums.StoreOrderHistoryType;
+import org.example.stockitbe.store.order.model.enums.StoreOrderStatus;
 import org.example.stockitbe.store.inbound.model.StoreInboundStatus;
 import org.example.stockitbe.store.order.model.entity.StoreOrderHeader;
 import org.example.stockitbe.store.order.model.entity.StoreOrderItem;
