+++ src/main/java/org/example/stockitbe/store/order/model/enums/StoreOrderHistoryType.java
@@ -1,6 +1,5 @@
-package org.example.stockitbe.store.order.model;
+package org.example.stockitbe.store.order.model.enums;
 
 public enum StoreOrderHistoryType {
     ORDER_STATUS
 }
-
