+++ src/main/java/org/example/stockitbe/store/order/model/enums/StoreOrderStatus.java
@@ -1,0 +1,8 @@
+package org.example.stockitbe.store.order.model.enums;
+
+public enum StoreOrderStatus {
+    REQUESTED,
+    APPROVED,
+    COMPLETED,
+    CANCELLED
+}
